Move getPersonInfo to user-utils
