+++ src/ggrc-client/js/components/related-objects/proposals/related-proposals-item.js
@@ -11,7 +11,7 @@ import '../../diff/instance-gca-diff';
 import '../../diff/instance-mapping-fields-diff';
 import '../../diff/instance-list-fields-diff';
 import template from './templates/related-proposals-item.mustache';
-import {getPersonInfo} from '../../../plugins/ggrc_utils';
+import {getPersonInfo} from '../../../plugins/utils/user-utils';
 import {getFormattedLocalDate} from '../../../plugins/utils/date-utils';
 const tag = 'related-proposals-item';
