+++ src/ggrc-client/js/components/related-objects/revisions/related-revisions-item.js
@@ -9,7 +9,7 @@ import '../../diff/instance-gca-diff';
 import '../../diff/instance-mapping-fields-diff';
 import '../../diff/instance-list-fields-diff';
 import '../../revision-history/restored-revision-comparer-config';
-import {getPersonInfo} from '../../../plugins/ggrc_utils';
+import {getPersonInfo} from '../../../plugins/utils/user-utils';
 import template from './templates/related-revisions-item.mustache';
 const tag = 'related-revisions-item';
